Move buffers to a separate header
