+++ include/RED4ext/NativeTypes-inl.hpp
@@ -288,79 +288,3 @@ RED4EXT_INLINE bool RED4ext::Variant::CanBeInlined(const RED4ext::CBaseRTTIType*
 {
     return aType->GetSize() <= InlineSize && aType->GetAlignment() <= InlineAlignment;
 }
-
-RED4EXT_INLINE RED4ext::RawBuffer::RawBuffer()
-    : data(nullptr)
-    , size(0)
-    , alignment(8)
-    , allocator{0}
-{
-}
-
-RED4EXT_INLINE RED4ext::RawBuffer::RawBuffer(void* aData, uint32_t aSize, uint32_t aAlignment)
-    : data(aData)
-    , size(aSize)
-    , alignment(aAlignment)
-    , allocator{0}
-{
-}
-
-RED4EXT_INLINE RED4ext::RawBuffer::~RawBuffer()
-{
-    if (data && allocator[0])
-    {
-        reinterpret_cast<RawBufferAllocator*>(&allocator)->Free(data);
-    }
-}
-
-RED4EXT_INLINE RED4ext::SharedPtr<RED4ext::DeferredDataBufferToken> RED4ext::DeferredDataBuffer::LoadAsync()
-{
-    using LoadBufferAsync_t = JobHandle* (*)(DeferredDataBuffer*, JobHandle*, int64_t);
-    static UniversalRelocFunc<LoadBufferAsync_t> func(Detail::AddressHashes::DeferredDataBuffer_LoadAsync);
-
-    JobHandle loadingJob;
-    func(this, &loadingJob, 0);
-
-    return MakeShared<DeferredDataBufferToken>(*this, loadingJob);
-}
-
-RED4EXT_INLINE RED4ext::SharedPtr<RED4ext::DeferredDataBufferRefToken> RED4ext::DeferredDataBuffer::LoadRefAsync()
-{
-    using LoadBufferRefAsync_t = void* (*)(DeferredDataBuffer*, SharedPtr<DeferredDataBufferRefToken>*, int8_t);
-    static UniversalRelocFunc<LoadBufferRefAsync_t> func(Detail::AddressHashes::DeferredDataBuffer_LoadRefAsync);
-
-    SharedPtr<DeferredDataBufferRefToken> token;
-    func(this, &token, 0);
-
-    return token;
-}
-
-RED4EXT_INLINE RED4ext::DeferredDataBufferToken::DeferredDataBufferToken(DeferredDataBuffer& aBuffer,
-                                                                         JobHandle& aJob) noexcept
-    : buffer(aBuffer)
-    , job(aJob)
-{
-}
-
-RED4EXT_INLINE void RED4ext::DeferredDataBufferToken::OnLoaded(LoadedCallback&& aCallback)
-{
-    JobQueue jobQueue;
-    jobQueue.Wait(job);
-    jobQueue.Dispatch([self = MakeShared<DeferredDataBufferToken>(*this), callback = std::move(aCallback)]()
-                      { callback(self->buffer); });
-}
-
-RED4EXT_INLINE RED4ext::DeferredDataBufferRefToken::~DeferredDataBufferRefToken()
-{
-    using DestructUnk28_t = void (*)(void**);
-    static UniversalRelocFunc<DestructUnk28_t> DestructUnk28(Detail::AddressHashes::ResourceToken_DestructUnk38);
-    DestructUnk28(&unk28);
-}
-
-RED4EXT_INLINE void RED4ext::DeferredDataBufferRefToken::OnLoaded(LoadedCallback&& aCallback)
-{
-    JobQueue jobQueue;
-    jobQueue.Wait(job);
-    jobQueue.Dispatch([self = self.Lock(), callback = std::move(aCallback)]()
-                      { callback(self); });
-}
