+++ include/RED4ext/NativeTypes.hpp
@@ -4,6 +4,7 @@
 #include <cstdint>
 #include <string_view>
 
+#include <RED4ext/Buffer.hpp>
 #include <RED4ext/CName.hpp>
 #include <RED4ext/CString.hpp>
 #include <RED4ext/Common.hpp>
@@ -166,112 +167,6 @@ struct Variant
     };
 };
 RED4EXT_ASSERT_SIZE(Variant, 0x18);
-
-struct RawBufferAllocator
-{
-    virtual void sub_00() = 0;                     // 08
-    virtual void Free(void* aData) = 0;            // 08
-    virtual Memory::IAllocator GetAllocator() = 0; // 10
-
-    void* unk08;
-};
-RED4EXT_ASSERT_SIZE(RawBufferAllocator, 0x10);
-
-struct RawBuffer
-{
-    using AllocatorType = Memory::EngineAllocator;
-
-    RawBuffer();
-    RawBuffer(void* aData, uint32_t aSize, uint32_t aAlignment = 8);
-    RawBuffer(const RawBuffer&) = delete;
-    RawBuffer(RawBuffer&&) = default;
-    ~RawBuffer();
-
-    void* data;            // 00
-    uint32_t size;         // 08
-    uint32_t alignment;    // 0C
-    uint64_t allocator[2]; // 10
-};
-RED4EXT_ASSERT_SIZE(RawBuffer, 0x20);
-
-struct DataBuffer
-{
-    RawBuffer buffer; // 00
-    uint64_t unk20;   // 20 - Pointer to something
-};
-RED4EXT_ASSERT_SIZE(DataBuffer, 0x28);
-
-struct SharedDataBuffer
-{
-    int64_t unk00; // 00
-};
-RED4EXT_ASSERT_SIZE(SharedDataBuffer, 0x8);
-
-struct DeferredDataBufferToken;
-struct DeferredDataBufferRefToken;
-
-enum class DeferredDataBufferState : uint8_t
-{
-    Unloaded = 0,
-    Loading = 1,
-    Loaded = 2,
-};
-
-struct DeferredDataBuffer
-{
-    SharedPtr<DeferredDataBufferToken> LoadAsync();
-    SharedPtr<DeferredDataBufferRefToken> LoadRefAsync();
-
-    RawBuffer temp;                // 00
-    SharedPtr<RawBuffer> raw;      // 20
-    void* unk30;                   // 30
-    uint8_t unk38;                 // 38
-    uint64_t unk40;                // 40
-    uint64_t unk48;                // 48
-    uint32_t unk50;                // 50
-    DeferredDataBufferState state; // 54
-    SharedMutex lock;              // 55
-    uint16_t unk56;                // 56
-};
-RED4EXT_ASSERT_SIZE(DeferredDataBuffer, 0x58);
-RED4EXT_ASSERT_OFFSET(DeferredDataBuffer, raw, 0x20);
-RED4EXT_ASSERT_OFFSET(DeferredDataBuffer, state, 0x54);
-RED4EXT_ASSERT_OFFSET(DeferredDataBuffer, lock, 0x55);
-
-struct DeferredDataBufferToken
-{
-    using AllocatorType = Memory::EngineAllocator;
-    using LoadedCallback = Callback<void (*)(DeferredDataBuffer&)>;
-
-    DeferredDataBufferToken(DeferredDataBuffer& aBuffer, JobHandle& aJob) noexcept;
-    DeferredDataBufferToken(const DeferredDataBufferToken&) = default;
-    DeferredDataBufferToken(DeferredDataBufferToken&&) = default;
-
-    void OnLoaded(LoadedCallback&& aCallback);
-
-    DeferredDataBuffer& buffer;
-    JobHandle job;
-};
-
-struct DeferredDataBufferRefToken
-{
-    using AllocatorType = Memory::EngineAllocator;
-    using LoadedCallback = Callback<void (*)(const SharedPtr<DeferredDataBufferRefToken>&)>;
-
-    ~DeferredDataBufferRefToken();
-
-    void OnLoaded(LoadedCallback&& aCallback);
-
-    WeakPtr<DeferredDataBufferRefToken> self; // 00
-    JobHandle job;                            // 10
-    SharedPtr<RawBuffer> raw;                 // 18
-    void* unk28;                              // 28 - SharedPtr.instance
-    void* unk30;                              // 30 - SharedPtr.refCount
-};
-RED4EXT_ASSERT_SIZE(DeferredDataBufferRefToken, 0x38);
-RED4EXT_ASSERT_OFFSET(DeferredDataBufferRefToken, job, 0x10);
-RED4EXT_ASSERT_OFFSET(DeferredDataBufferRefToken, raw, 0x18);
-RED4EXT_ASSERT_OFFSET(DeferredDataBufferRefToken, unk28, 0x28);
 
 struct gamedataLocKeyWrapper
 {
